pyproject.toml
[project]
name = "daily-flight-deck"
version = "0.1.0"
description = "SE Daily Flight Deck morning brief generator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "jinja2>=3.1",
  "openai>=1.40.0",
  "PyYAML>=6.0",
  "uvicorn>=0.30",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
